+++ source/Processing.Infrastructure/DomainEventDispatching/DomainEventsAccessor.cs
@@ -16,7 +16,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Processing.Domain.SeedWork;
-using Processing.Infrastructure.DataAccess;
+using Processing.Infrastructure.Configuration.DataAccess;
 
 namespace Processing.Infrastructure.DomainEventDispatching
 {
