+++ source/Processing.Infrastructure/InternalCommands/InternalCommandAccessor.cs
@@ -16,7 +16,7 @@ using System;
 using System.Collections.Immutable;
 using System.Linq;
 using System.Threading.Tasks;
-using Processing.Infrastructure.DataAccess;
+using Processing.Infrastructure.Configuration.DataAccess;
 
 namespace Processing.Infrastructure.InternalCommands
 {
