+++ source/Processing.Infrastructure/InternalCommands/InternalCommandProcessor.cs
@@ -15,7 +15,7 @@
 using System;
 using System.Threading.Tasks;
 using Processing.Domain.SeedWork;
-using Processing.Infrastructure.DataAccess;
+using Processing.Infrastructure.Configuration.DataAccess;
 
 namespace Processing.Infrastructure.InternalCommands
 {
