+++ source/Processing.Infrastructure/Outbox/OutboxManager.cs
@@ -14,7 +14,7 @@
 
 using System.Linq;
 using Processing.Domain.SeedWork;
-using Processing.Infrastructure.DataAccess;
+using Processing.Infrastructure.Configuration.DataAccess;
 
 namespace Processing.Infrastructure.Outbox
 {
