+++ source/Processing.IntegrationTests/Application/ChangeOfSupplier/Processing/Commands/ChangeSupplierTests.cs
@@ -26,7 +26,7 @@ using Processing.Application.ChangeOfSupplier.Processing.MeteringPointDetails;
 using Processing.Domain.Consumers;
 using Processing.Domain.EnergySuppliers;
 using Processing.Domain.MeteringPoints;
-using Processing.Infrastructure.DataAccess;
+using Processing.Infrastructure.Configuration.DataAccess;
 using Processing.Infrastructure.Integration.IntegrationEvents.FutureEnergySupplierChangeRegistered;
 using Xunit;
 using Xunit.Categories;
