+++ source/Processing.IntegrationTests/Application/TestHost.cs
@@ -52,13 +52,13 @@ using Processing.Domain.MeteringPoints.Events;
 using Processing.Domain.SeedWork;
 using Processing.Infrastructure.BusinessRequestProcessing;
 using Processing.Infrastructure.BusinessRequestProcessing.Pipeline;
+using Processing.Infrastructure.Configuration.DataAccess;
+using Processing.Infrastructure.Configuration.DataAccess.AccountingPoints;
+using Processing.Infrastructure.Configuration.DataAccess.Consumers;
+using Processing.Infrastructure.Configuration.DataAccess.EnergySuppliers;
+using Processing.Infrastructure.Configuration.DataAccess.ProcessManagers;
 using Processing.Infrastructure.ContainerExtensions;
 using Processing.Infrastructure.Correlation;
-using Processing.Infrastructure.DataAccess;
-using Processing.Infrastructure.DataAccess.AccountingPoints;
-using Processing.Infrastructure.DataAccess.Consumers;
-using Processing.Infrastructure.DataAccess.EnergySuppliers;
-using Processing.Infrastructure.DataAccess.ProcessManagers;
 using Processing.Infrastructure.DomainEventDispatching;
 using Processing.Infrastructure.EDI;
 using Processing.Infrastructure.EDI.ChangeOfSupplier;
